+++ CHANGELOG.md
@@ -9,3 +9,4 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ### Added
 - "Show # of Files" command
 - Include/Exclude configuration options
+- Option to use .gitignore for excluding files